Update features.py
When calling atom_to_feature_vector() the atomic number is not correct. The safe_index()  function is used which gives the index of the atomic number in the list 'possible_atomic_num_list' which currently starts at 1. This means that when looking for an atomic number of, say, 8 we get 7 instead. I propose that the list start at 0 to rectify this issue. 
This may lead to an issue if atom.GetAtomicNum ever returns 0 but I am unsure how likely this is.